python_ML/requirements.txt
pandas>=2.0.3
numpy>=1.24.3
scikit-learn>=1.3.0
pickle-mixin>=1.0.2
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
matplotlib>=3.9.4
requests>=2.28.1

# FastAPI + production-ready ASGI server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
